feat(dataconnect): restructure existing transport code into new file structure (#9603)
* restructure existing transport code into new file structure

* remove DataConnectResponse from public API

* update copyright year
