+++ packages/data-connect/src/network/rest/RestTransport.ts
@@ -15,20 +15,20 @@
  * limitations under the License.
  */
 
-import { DataConnectOptions, TransportOptions } from '../../api/DataConnect';
-import { AppCheckTokenProvider } from '../../core/AppCheckTokenProvider';
-import { DataConnectError, Code } from '../../core/error';
-import { AuthTokenProvider } from '../../core/FirebaseAuthProvider';
-import { logDebug } from '../../logger';
-import { addToken, urlBuilder } from '../../util/url';
-import { dcFetch } from '../fetch';
-
 import {
   CallerSdkType,
   CallerSdkTypeEnum,
   DataConnectResponse,
   DataConnectTransport
-} from '.';
+} from '..';
+import { DataConnectOptions, TransportOptions } from '../../api/DataConnect';
+import { AppCheckTokenProvider } from '../../core/AppCheckTokenProvider';
+import { Code, DataConnectError } from '../../core/error';
+import { AuthTokenProvider } from '../../core/FirebaseAuthProvider';
+import { logDebug } from '../../logger';
+import { addToken, urlBuilder } from '../../util/url';
+
+import { dcFetch } from './fetch';
 
 export class RESTTransport implements DataConnectTransport {
   private _host = '';
