+++ packages/data-connect/test/unit/gmpid.test.ts
@@ -21,7 +21,7 @@ import * as sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 
 import { DataConnect, executeQuery, getDataConnect, queryRef } from '../../src';
-import { initializeFetch } from '../../src/network/fetch';
+import { initializeFetch } from '../../src/network/rest';
 
 use(sinonChai);
 const json = {
